requirements.txt
faster-coco-eval>=1.6.5
PyYAML
tensorboard
scipy
calflops
thop
transformers
pytorch_wavelets==1.3.0
timm==1.0.7
tidecv
einops
prettytable
pycocotools==2.0.8